Переименовал проект под Lesson_8.
