+++ .idea/.name
@@ -1,1 +1,1 @@
-Lesson_7_Notes
+Lesson_8_Notes
