+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "30.0.3"
 
     defaultConfig {
-        applicationId "Android_Intro.Lesson_7_Notes"
+        applicationId "Android_Intro.Lesson_8_Notes"
         minSdkVersion 24
         targetSdkVersion 30
         versionCode 1
