+++ app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="Android_Intro.Lesson_7_Notes">
+    package="Android_Intro.Lesson_8_Notes">
 
     <application
         android:allowBackup="true"
@@ -7,8 +7,8 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Lesson_7_Notes.NoActionBar">
-        <activity android:name="Android_Intro.Lesson_7_Notes.MainActivity">
+        android:theme="@style/Theme.Lesson_8_Notes.NoActionBar">
+        <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
