+++ app/src/main/java/Android_Intro/Lesson_8_Notes/MainActivity.java
@@ -1,4 +1,4 @@
-package Android_Intro.Lesson_7_Notes;
+package Android_Intro.Lesson_8_Notes;
 
 
 import androidx.appcompat.app.AppCompatActivity;
