+++ app/src/main/res/layout/fragment_note_screen.xml
@@ -11,7 +11,7 @@
         android:id="@+id/menu_toolbar_fragment_note"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:theme="@style/Theme.Lesson_7_Notes.AppBarOverlay">
+        android:theme="@style/Theme.Lesson_8_Notes.AppBarOverlay">
 
         <com.google.android.material.appbar.MaterialToolbar
             android:id="@+id/toolbar_main_fragment"
@@ -19,7 +19,7 @@
             android:layout_height="?attr/actionBarSize"
             android:background="?attr/colorPrimary"
             android:backgroundTint="@color/toolbar_bg"
-            app:popupTheme="@style/Theme.Lesson_7_Notes.PopupOverlay" />
+            app:popupTheme="@style/Theme.Lesson_8_Notes.PopupOverlay" />
 
     </com.google.android.material.appbar.AppBarLayout>
