+++ settings.gradle
@@ -1,2 +1,2 @@
 include ':app'
-rootProject.name = "Lesson_7_Notes"
+rootProject.name = "Lesson_8_Notes"
